fix: create reusable function for authentication
