+++ src/routes/ProtectedRoute.tsx
@@ -1,22 +1,9 @@
+import { getAuthInfo } from '@/util/helpers'
 import { ProtectedRouteProps } from '@/util/interfaces'
 import { Navigate, Outlet } from 'react-router-dom'
 
 export function ProtectedRoute({ allowedRoles }: ProtectedRouteProps) {
-    const isAuthenticated = !!localStorage.getItem('token')
-
-    let userRole = null
-    try {
-        const user = localStorage.getItem('user')
-        if (user) {
-            const userData = JSON.parse(user)
-            userRole = userData.roles
-        }
-    } catch (error) {
-        console.error('Failed to parse user data:', error)
-        localStorage.removeItem('user')
-        localStorage.removeItem('token')
-        return <Navigate to="/login" replace />
-    }
+    const { isAuthenticated, userRole } = getAuthInfo()
 
     if (!isAuthenticated) {
         return <Navigate to="/login" replace />
